+++ thoth-app/src/component/work.rs
@@ -35,6 +35,7 @@ use crate::component::delete_dialogue::ConfirmDeleteComponent;
 use crate::component::fundings_form::FundingsFormComponent;
 use crate::component::issues_form::IssuesFormComponent;
 use crate::component::languages_form::LanguagesFormComponent;
+use crate::component::new_chapter::NewChapterComponent;
 use crate::component::publications_form::PublicationsFormComponent;
 use crate::component::related_works_form::RelatedWorksFormComponent;
 use crate::component::subjects_form::SubjectsFormComponent;
@@ -814,6 +815,22 @@ impl Component for WorkComponent {
                             current_user=self.props.current_user.clone()
                             update_relations=self.link.callback(Msg::UpdateRelatedWorks)
                         />
+                        {
+                            if self.work.work_type == WorkType::Monograph
+                                || self.work.work_type == WorkType::EditedBook
+                                || self.work.work_type == WorkType::Textbook {
+                                html! {
+                                    // Convenience button for adding Chapter relations with inherited values
+                                    <NewChapterComponent
+                                        work=self.work.clone()
+                                        relations=self.work.relations.clone()
+                                        update_relations=self.link.callback(Msg::UpdateRelatedWorks)
+                                    />
+                                }
+                            } else {
+                                html! {}
+                            }
+                        }
                         <ContributionsFormComponent
                             contributions=self.work.contributions.clone()
                             work_id=self.work.work_id